Move serializer in its own folder
